Handle source being updated before the animation has finished
Adding a flag ongoingAnimation to know if we can call setNewSource or not yet.

If we can't call it (ongoing animation in progress), then we give that reference to setNextSource.

setNextSource is called after the animation is ready.

If the source is changed multiple times before an animation if finished, only the last source will be animated, previous ones being skipped - no need to animate them all if the source changed so fast.

Using a callback instead of a setTimeout for performance purposes and to skip sources in between and animate only the last one.